toggling components functionality.
we can exit the gameDetail by clicking on the background space that we leave between our gameDetail and game component.
Also if we have our gameDetail open we will not be able to scroll our game component as we set the overflow to hidden in game, but then when we come back we set the overflow to auto. we use useNavigate to navigate from gameDetail to game by clicking on the background space.