implementing nav searching.
rendering out a search componenet when we serach for a game in our Nav bar. we made a new API call for searched games and then use conditional rendering where if the lenght of searched games is more than 0 (which means we are searching for a game) then we render out the searched game component.